add initial layers configuration
The config object may now hold a "layers" entry, expected to be an array
of objects {id::string, visibility::boolean, transparency::0-100}. Those
objects will be used to create the layers from the services.

createMap will fall back to the lgv default layer 453, as long as the
lgv services.json is used.

MPAPI-1