add search function
In the folder src/searchAddress a search function is exported that uses
the HH gazetteer to retrieve information for various search types. Every
search result is parsed to a default format containing name, geometry,
and all retrieved properties.

Searching for house numbers is a special case: It only works when also
searching for streets, and all house numbers for all found streets are
retrieved.

There is currently no GUI element to use this function with. You can run
the search function when running the example within your browser's
console.

MPAPI-9